% Auto-generated: do not edit by hand
\name{dashGridCol}

\alias{dashGridCol}

\title{DashGridCol component}

\description{

}

\usage{
dashGridCol(children=NULL, size=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{size}{List of numerics. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

