% Auto-generated: do not edit by hand
\name{dashGridLayout}

\alias{dashGridLayout}

\title{DashGridLayout component}

\description{

}

\usage{
dashGridLayout(children=NULL, id=NULL, addItem=NULL, className=NULL,
cols=NULL, compactType=NULL, currentLayout=NULL,
itemCount=NULL, newItemTemplate=NULL, onLayoutChange=NULL,
rowHeight=NULL, showRemoveButton=NULL,
showResizeHandles=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{addItem}{Logical. }

\item{className}{Character. }

\item{cols}{Named list. }

\item{compactType}{A value equal to: 'vertical', 'horizontal', null. }

\item{currentLayout}{List of lists containing elements 'i', 'x', 'y', 'w', 'h'.
those elements have the following types:
  - i (character; optional)
  - x (numeric; optional)
  - y (numeric; optional)
  - w (numeric; optional)
  - h (numeric; optional)s. }

\item{itemCount}{Numeric. }

\item{newItemTemplate}{A list of or a singular dash component, string or number. }

\item{onLayoutChange}{}

\item{rowHeight}{Numeric. }

\item{showRemoveButton}{Logical. }

\item{showResizeHandles}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

