% Auto-generated: do not edit by hand
\name{draggableWrapper}

\alias{draggableWrapper}

\title{DraggableWrapper component}

\description{

}

\usage{
draggableWrapper(children=NULL, id=NULL, handleBackground=NULL,
handleColor=NULL, handleText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The content to be wrapped and made draggable.}

\item{id}{Character. A unique identifier for the DraggableWrapper component.}

\item{handleBackground}{Character. The background color of the drag handle.}

\item{handleColor}{Character. The text color of the drag handle.}

\item{handleText}{Character. The text to display in the drag handle.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

