% Auto-generated: do not edit by hand
\name{dashGridLayout}

\alias{dashGridLayout}

\title{DashGridLayout component}

\description{

}

\usage{
dashGridLayout(id=NULL, addItem=NULL, breakpointData=NULL,
breakpoints=NULL, className=NULL, cols=NULL,
compactType=NULL, currentLayout=NULL, itemCount=NULL,
itemLayout=NULL, items=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, rowHeight=NULL,
showRemoveButton=NULL, showResizeHandles=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{addItem}{Logical. }

\item{breakpointData}{Lists containing elements 'newbreakpoint', 'newcols'.
those elements have the following types:
  - newbreakpoint (character; optional)
  - newcols (numeric; optional). }

\item{breakpoints}{Lists containing elements 'lg', 'md', 'sm', 'xs', 'xxs'.
those elements have the following types:
  - lg (numeric; optional)
  - md (numeric; optional)
  - sm (numeric; optional)
  - xs (numeric; optional)
  - xxs (numeric; optional). }

\item{className}{Character. }

\item{cols}{Named list. }

\item{compactType}{A value equal to: 'vertical', 'horizontal', null. }

\item{currentLayout}{List of lists containing elements 'i', 'x', 'y', 'w', 'h'.
those elements have the following types:
  - i (character; optional)
  - x (numeric; optional)
  - y (numeric; optional)
  - w (numeric; optional)
  - h (numeric; optional)s. }

\item{itemCount}{Numeric. }

\item{itemLayout}{List of lists containing elements 'i', 'x', 'y', 'w', 'h'.
those elements have the following types:
  - i (character; optional)
  - x (numeric; optional)
  - y (numeric; optional)
  - w (numeric; optional)
  - h (numeric; optional)s. }

\item{items}{List of a list of or a singular dash component, string or numbers. }

\item{persisted_props}{Unnamed list. }

\item{persistence}{Logical. }

\item{persistence_type}{A value equal to: 'local', 'memory', 'session'. }

\item{rowHeight}{Numeric. }

\item{showRemoveButton}{Logical. }

\item{showResizeHandles}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

