% Auto-generated: do not edit by hand
\name{dashGridLayout}

\alias{dashGridLayout}

\title{DashGridLayout component}

\description{

}

\usage{
dashGridLayout(children=NULL, id=NULL, allowAddItem=NULL, className=NULL,
cols=NULL, compactType=NULL, itemCount=NULL,
newItemTemplate=NULL, onLayoutChange=NULL, rowHeight=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{allowAddItem}{Logical. }

\item{className}{Character. }

\item{cols}{Named list. }

\item{compactType}{A value equal to: 'vertical', 'horizontal', null. }

\item{itemCount}{Numeric. }

\item{newItemTemplate}{A list of or a singular dash component, string or number. }

\item{onLayoutChange}{}

\item{rowHeight}{Numeric. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

