% Auto-generated: do not edit by hand
\name{draggableWrapper}

\alias{draggableWrapper}

\title{DraggableWrapper component}

\description{

}

\usage{
draggableWrapper(children=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

